+++ apps/dsp-app/src/app/user/create-user-page/create-user-dialog.component.ts
@@ -15,7 +15,7 @@ import { UserForm } from '../user-form/user-form.type';
     <app-user-form [user]="user" (afterFormInit)="form = $event" />
 
     <div mat-dialog-actions align="end">
-      <button color="primary" mat-button mat-dialog-close>{{ 'appLabels.form.action.cancel' | translate }}</button>
+      <button color="primary" mat-button mat-dialog-close>{{ 'form.action.cancel' | translate }}</button>
       <button
         mat-raised-button
         color="primary"
@@ -23,7 +23,7 @@ import { UserForm } from '../user-form/user-form.type';
         [isLoading]="isLoading$ | async"
         [disabled]="!form?.valid || (isLoading$ | async)"
         (click)="createUser()">
-        {{ 'appLabels.form.action.submit' | translate }}
+        {{ 'form.action.submit' | translate }}
       </button>
     </div>
   `,
@@ -56,7 +56,11 @@ export class CreateUserDialogComponent {
     user.systemAdmin = this.form.controls.systemAdmin.value;
     user.status = true;
 
-    this._store.dispatch(new CreateUserAction(user, this._projectService.uuidToIri(this.projectUuId)));
+    const projectIri =
+      typeof this.projectUuId === 'string' && this.projectUuId.length > 0
+        ? this._projectService.uuidToIri(this.projectUuId)
+        : undefined;
+    this._store.dispatch(new CreateUserAction(user, projectIri));
     this._actions$.pipe(ofActionSuccessful(CreateUserAction), take(1)).subscribe(() => {
       this._dialogRef.close();
       const enrolled = this.projectUuId ? ' and added the user to the project' : '';