+++ apps/dsp-app/src/app/user/user.component.ts
@@ -3,10 +3,10 @@ import { Title } from '@angular/platform-browser';
 import { ActivatedRoute } from '@angular/router';
 import { RouteConstants } from '@dasch-swiss/vre/shared/app-config';
 import { UserSelectors } from '@dasch-swiss/vre/shared/app-state';
+import { TranslateService } from '@ngx-translate/core';
 import { Select } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { AppGlobal } from '../app-global';
 import { MenuItem } from '../main/declarations/menu-item';
 
 @Component({
@@ -25,7 +25,20 @@ export class UserComponent implements OnDestroy {
   // for the sidenav
   open = true;
 
-  navigation: MenuItem[] = AppGlobal.userNav;
+  navigation: MenuItem[] = [
+    {
+      label: this._translateService.instant('user.navigation.myProjects'),
+      shortLabel: this._translateService.instant('user.navigation.myProjects'),
+      route: RouteConstants.projectsRelative,
+      icon: 'assignment',
+    },
+    {
+      label: this._translateService.instant('user.navigation.myAccount'),
+      shortLabel: this._translateService.instant('user.navigation.myAccount'),
+      route: RouteConstants.userAccountRelative,
+      icon: 'settings',
+    },
+  ];
 
   @Select(UserSelectors.username) username$: Observable<string>;
 
@@ -33,7 +46,8 @@ export class UserComponent implements OnDestroy {
 
   constructor(
     private _route: ActivatedRoute,
-    private _titleService: Title
+    private _titleService: Title,
+    private _translateService: TranslateService
   ) {
     // get the activated route; we need it for the viewer switch
     this.route = this._route.pathFromRoot[1].snapshot.url[0].path;