+++ libs/vre/shared/app-resource-properties/src/lib/properties-display-incoming-link.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { ReadResource, ReadResourceSequence } from '@dasch-swiss/dsp-js';
 import { IncomingService } from '@dasch-swiss/vre/shared/app-common-to-move';
-import { Observable } from 'rxjs';
-import { map, shareReplay } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { expand, map, reduce, takeWhile } from 'rxjs/operators';
 import { IncomingOrStandoffLink } from './incoming-link.interface';
 import { sortByKeys } from './sortByKeys';
 
@@ -11,7 +11,28 @@ export class PropertiesDisplayIncomingLinkService {
   constructor(private _incomingService: IncomingService) {}
 
   searchIncomingLinks$ = (resourceId: string, offset: number) =>
-    this._incomingService.getIncomingLinksForResource(resourceId, offset).pipe(shareReplay());
+    this._incomingService.getIncomingLinksForResource(resourceId, offset);
+
+  getIncomingLinksRecursively$(resourceId: string, offset: number = 0): Observable<IncomingOrStandoffLink[]> {
+    return this._incomingService.getIncomingLinksForResource(resourceId, offset).pipe(
+      expand(sequence => {
+        if ((sequence as ReadResourceSequence).mayHaveMoreResults === false) {
+          return of(sequence as ReadResourceSequence);
+        }
+
+        return this._incomingService.getIncomingLinksForResource(
+          resourceId,
+          ++offset
+        ) as Observable<ReadResourceSequence>;
+      }),
+      takeWhile(response => response.resources.length > 0 && response.mayHaveMoreResults === true, true),
+      reduce((all: ReadResource[], data) => all.concat(data.resources), []),
+      map(incomingResources => {
+        const incomingLinks = incomingResources.map(resource => this._createIncomingOrStandoffLink(resource));
+        return sortByKeys(incomingLinks, ['project', 'label']);
+      })
+    );
+  }
 
   getIncomingLinks$(searchResults$: Observable<ReadResourceSequence>): Observable<IncomingOrStandoffLink[]> {
     return searchResults$.pipe(