+++ libs/vre/shared/app-string-literal/src/lib/multi-language-form.service.ts
@@ -1,8 +1,9 @@
 import { ChangeDetectorRef, Injectable } from '@angular/core';
 import { FormBuilder, ValidatorFn } from '@angular/forms';
+import { AvailableLanguages } from '@dasch-swiss/vre/shared/app-config';
 import { UserSelectors } from '@dasch-swiss/vre/shared/app-state';
 import { Store } from '@ngxs/store';
-import { DaschLanguage, isDaschLanguage } from './dash-language.type';
+import { isDaschLanguage } from './dash-language.type';
 import { MultiLanguageFormArray } from './multi-language-form-array.type';
 
 /** Component Provider used in combination with
@@ -11,7 +12,7 @@ import { MultiLanguageFormArray } from './multi-language-form-array.type';
  */
 @Injectable()
 export class MultiLanguageFormService {
-  readonly availableLanguages: DaschLanguage[] = ['de', 'fr', 'it', 'en', 'rm'];
+  readonly availableLanguages: string[] = AvailableLanguages.map(lang => lang.language!);
   selectedLanguageIndex: number;
   formArray: MultiLanguageFormArray;
   validators: ValidatorFn[];